+++ extensions/qrcode-generator/CHANGELOG.md
@@ -4,3 +4,7 @@
 
 - Optimized the code
 - Added screenshots
+
+## [Update] - 2024-06-22
+
+- Added feature to copy generated QR image (from clipboard link) back to the clipboard